Remove old DateFormatter and use only one across all SDKs
