--- stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/DateConverter.kt
@@ -1,15 +1,0 @@
-package io.getstream.chat.android.ui.utils
-
-import org.threeten.bp.Instant
-import org.threeten.bp.LocalDateTime
-import org.threeten.bp.ZoneId
-import java.util.Date
-
-internal object DateConverter {
-
-    fun toLocalDateTime(date: Date): LocalDateTime {
-        return Instant.ofEpochMilli(date.time)
-            .atZone(ZoneId.systemDefault())
-            .toLocalDateTime()
-    }
-}
