+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/utils/DateFormatter.kt
@@ -5,17 +5,17 @@ import org.threeten.bp.LocalDateTime
 import java.util.Date
 
 public interface DateFormatter {
-    public fun format(localDateTime: LocalDateTime?): String
+    public fun formatDate(localDateTime: LocalDateTime?): String
 
     public companion object {
         @JvmStatic
-        public fun getDefault(context: Context): DateFormatter = DefaultDateFormatter(context)
+        public fun from(context: Context): DateFormatter = DefaultDateFormatter(context)
     }
 }
 
 /**
  * Extension to be able to format objects of the deprecated [Date] type.
  */
-internal fun DateFormatter.format(date: Date?): String {
-    return format(date?.let(DateConverter::toLocalDateTime))
+public fun DateFormatter.formatDate(date: Date?): String {
+    return formatDate(date?.let(DateConverter::toLocalDateTime))
 }
